fdt: Move to setuptools
The distutils package is deprecated. The upstream libfdt repo uses
setuptools for building the pylibfdt module, so bring in that code,
suitably modified for U-Boot. Also bring in the README.

The modifications include setting the version correctly, making use of
the environment variables provided by the Makefile and various tweaks
to the directories.

Note that the version omits the minus character at the start of
EXTRAVERSION, since this creates a warning. The build is really just used
within U-Boot itself, so it doesn't matter too much if the version matches
upstream, or exactly matches U-Boot.

Series-to: u-boot
Series-cc: trini
Cover-letter:
fdt: Fix up tests and build warnings on newer distros
At present the code coverage for test_fdt is not working. This series
fixes that and takes the opportunity to fix the pylint warnings in that
file.

Some newer Linux distributions are producing warnings about using
distutils, so this series moves things over to setuptools.
END

Signed-off-by: Simon Glass <sjg@chromium.org>