sandbox: Add a test for SCSI
Add a simple uclass test for SCSI. It reads the partition table from a
disk image and checks that it looks correct.

Series-to: u-boot
Series-links: 315656
Cover-letter:
scsi: Convert sandbox SCSI to driver model
This series moves sandbox over to use driver model for SCSI. It takes the
opportunity to use common code with the USB flash driver.

This is mostly refactoring, but there is a small amount of new code,
notably a test which checks that data can be read from a SCSI device.
END
Signed-off-by: Simon Glass <sjg@chromium.org>