vbe: Add a test for VBE device tree fixups
When a FIT includes some OS requests, U-Boot should process these and add
the requested info to corresponding subnodes of the /chosen node. Add a
pytest for this, which sets up the FIT, runs bootm and then uses a C
unit test to check that everything looks OK.

The test needs to run on sandbox_flattree since we don't support
device tree fixups on sandbox (live tree) yet. So enable BOOTMETH_VBE and
disable bootflow_system(), since EFI is not supported on
sandbox_flattree.

Add a link to the initial documentation.

Series-to: u-boot
Cover-letter:
vbe: Support device tree fixups for OS requests
VBE provides the ability for an OS to request that information be passed
to it when it is booted. This is added into the /chosen node, in addition
to things like the bootargs for Linux, for example.

VBE's OS requests are intended to replace the need for the EFI 'boot-time
services'. This works via a 'stub' which runs before Linux, collects the
information from U-Boot, writes it to the device tree (mostly) and then
jumps to Linux with the updated device tree.

Rather than just jumping into Linux and waiting for it to request things
from U-Boot, we can look at the requests in the FIT and process them
before jumping to Linux. This is simpler and easier to test. It is also
more deterministic, since we can tell whether we might lack something
needed by Linux, before jumping to it.

This series adds initial support for OS requests, with just a few simple
ones provided. Further work will expand these out.
END