dm: blk: mmc: Tidy up some Makefile rules for SPL
Use the correct SPL_TPL_ variable so that these features can be enabled in
TPL and VPL as needed.

Note that the default for TPL is y since some boards boot via MMC, so
this addds 137 bytes on rockchip firefly-rk3288 for example.

Other boards which enable BLK in TPL are:

   chromebook_coral
   chromebook_samus_tpl
   sandbox_vpl

Signed-off-by: Simon Glass <sjg@chromium.org>