vbe: Add Kconfig options for VPL
Enable the various features needed in VPL, by adding Kconfig options.

Update the defconfig for sandbox_vpl so that the build for each phase
includes what is needed. Drop LZMA for now and make sure partition support
is omitted in SPL, since it is not needed.

Fix up the sandbox_vpl setings so that it still builds.

Signed-off-by: Simon Glass <sjg@chromium.org>