vbe: Add a test for the VBE flow into U-Boot proper
Add a test which checks that VBE boots correctly from TPL through to
U-Boot proper. This works by attaching the image.bin file to mmc3 so
that it is accessible to sandbox_vpl. We don't want to include this
in the generic sandbox build, since that image is of zero size and
cannot be mapped with mmap().

Series-to: u-boot
Series-cc: trini
Cover-letter:
vbe: Implement the full firmware flow
This series provides an implementation of VBE from TPL through to U-Boot
proper, using VBE to load the relevant firmware stages. It builds a single
image.bin file containing all the phases:

   TPL - initial phase, loads VPL using binman symbols
   VPL - main firmware phase, loads SPL using VBE parameters
   SPL - loads U-Boot proper using VBE parameters
   U-Boot - final firmware phase, where OS booting is processed

This series does not include the OS-booting phase. The previous series
included support for OS requests. The rest of it will be the subject of a
future series.

The implementation is entirely handled by sandbox. It should be possible
to enable this on a real board without much effort, but that is also the
subject of a future series.

Apart from future work, there is other one piece to be resolved. The
concept of a 'phase' is introduced into FIT configurations. Perhaps
instead this should be attached to an image, not a configuration,
since VPL should select the (only) configuration once, passing that
through to SPL to use as well? This would drop the need to verify
signatures, etc. in SPL.

This needs some more thought and investigation. Another idea is to use the
loadables feature with a list of images to load, e.g. loadables-vpl and
loadables-spl for the two phases.
END

Signed-off-by: Simon Glass <sjg@chromium.org>