buildman: Detect binman reporting missing blobs
Buildman should consider a build as a success (with warnings) if missing
blobs have been dealt with by binman, even though buildman itself returns
and error code overall. This is how other warnings are dealt with.

We cannot easily access the 103 exit code, so detect the problem in the
output.

With this change, missing blobs result in an exit code of 101, although
they still indicate failure.

Series-to: u-boot
Series-cc: trini
Series-cc: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Cover-letter:
Makefile: Deal with missing blobs consistently
Missing blobs should cause the build (with make) to fail, but at present
success is returned. This is because binman currently produces an exit
code of 0 in this case.

Of course this is not correct, since the images cannot actually be used.

This series fixes that and adjusts buildman to deal sensibly with the
situation.

It also includes a buildman patch to deal with N: entries in the
MAINTAINER files and a few other minor niggles noticed along the way.
END

Signed-off-by: Simon Glass <sjg@chromium.org>