buildman: Add --allow-missing flag to allow missing blobs
Add a new flag to buildman so that we will in turn pass
BINMAN_ALLOW_MISSING=1 to 'make'. Make use of this flag in CI.

Allow the settings file to control this.

Series-to: u-boot
Series-cc: trini
Series-links: 326991 3:322306
Series-version: 4
Cover-letter:
buildman: Correct various issues with missing blobs
This series attempts to tidy up the problem of missing blobs, making
U-Boot default to failing the build, but providing ways to configure this
so that buildman can still be used for build testing by people who don't
have the blobs or don't care about them.

This is based on discussion in [1] as well as some experimentation.

It also moves the buildman docs to rST and tidies up some toolchain
settings and documentation.

[1] https://patchwork.ozlabs.org/project/uboot/patch/20221010151831.3376759-2-trini@konsulko.com/
END

Series-changes: 3
- Add tests docs and a settings-file option

Cc: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Cc: Simon Glass <sjg@chromium.org>
Signed-off-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Simon Glass <sjg@chromium.org>