binman: Allow writing section contents to a file
At present only the image (which is a section) has a filename. Move this
implementation to the entry_Section class so that any section can have a
filename. With this, the section data is written to a file.

This allows parts of an image to be written, along with the entire image.

Make a note that this can be used to include the contents of a section in
one image in another (later) image.

Series-changes: 5
- Update commit message to mention using parts of one image in another

Signed-off-by: Simon Glass <sjg@chromium.org>