binman: Support new op-tee binary format
OP-TEE has a format with a binary header that can be used instead of the
ELF file. With newer versions of OP-TEE this may be required on some
platforms.

Add support for this in binman. First, add a method to obtain the ELF
sections from an entry, then use that in the FIT support. We then end up
with the ability to support both types of OP-TEE files, depending on which
one is passed in with the entry argument (TEE=xxx in the U-Boot build).

Series-changes: 6
- Update op-tee to support new v1 binary header

Series-changes: 7
- Correct missing test coverage

Series-changes: 8
- Fix 'anb', 'provide' and 'second'  typos
- Don't silence the return code from tee_os.ObtainContents()
- Rename is_optee_bin()
- Improve various comments

Signed-off-by: Simon Glass <sjg@chromium.org>