binman: Support overlaying entries inside blobs
Where a binary contains multiple sections it is sometimes useful to make
these visible in the fdtmap, so that invidual parts can be accessed. For
example, there may be a version string at a fixed position in the binary.

In addition, it is sometimes useful to pad sections with given data,
rather than just 0 or 0xff bytes.

We can implement both of these with a 'section-fill' feature. Add support
a subnode of this name, which provides the data to fill a section. The
actual size of the section is set by its contents (as usual), but any
space between the entries is filled up by the contents of the fill data.

Signed-off-by: Simon Glass <sjg@chromium.org>