timer: Tidy up use of notrace
Tracing is typically enabled by the time driver model starts up, so ther
is no point in adding a 'notrace' to the timer-init function. However, we
do need to be able to access the timer's private data when reading the
timer, so add it to the core function needed for that.

Update the function's documentation while we are here.

Signed-off-by: Simon Glass <sjg@chromium.org>