env: Make nvedit condition on CONFIG_ENV_SUPPORT
We cannot use the environment when it is not enabled.

This works OK at present, but with the updated kconfig system it hits a
strange corner case. In this case, CONFIG_SPL_ENV_IS_NOWHERE shows up as
CONFIG_ENV_IS_NOWHERE in SPL which trips the check at the top of the file.

Fix this by avoiding what is anyway a useless compilation.

Signed-off-by: Simon Glass <sjg@chromium.org>