bootstd: Add some default filesystems and commands
We need to support a basic set of filesystems for booting to work in most
cases. Add these in, letting the board disable them (for space reasons) if
desired.

This does enable the commands as well as the functionality, even though
bootstd is quite happen to use ext4 without the ext4 command. Further work
would be needed to disintangle this and reduce code size.

Signed-off-by: Simon Glass <sjg@chromium.org>