Correct SPL uses of ZYNQMP_FIRMWARE
This converts 2 usages of this option to the non-SPL form, since there is
no SPL_ZYNQMP_FIRMWARE defined in Kconfig

Series-to: u-boot
Series-cc: trini
Cover-letter:
Kconfig: More cleanup of CONFIG options
Quite a few CONFIG options without an SPL equivalent are used as if they
have one. For example, we see CONFIG_IS_ENABLED(IOMMU) when there is no
SPL_IOMMU Kconfig option.

This series adds new features to moveconfig to detect and remove these
problems. It includes additions to the Kconfig for options which are
needed as well as dropping unwanted us of CONFIG_IS_ENABLED, etc.

This follows on from the previous series[1]

It is available at u-boot-dm/splb-working

[1] https://patchwork.ozlabs.org/project/uboot/list/?series=339004
END

Signed-off-by: Simon Glass <sjg@chromium.org>