input: Only reset the keyboard when running bare metal
If U-Boot is not the first-stage bootloader we should not init the
keyboard, since it has already been done. Check for this.

This is needed for brya when starting from coreboot, depending on how
it was reset. In any case, it sems reasonable to avoid a reset if a
previous stage has skipped it.

Signed-off-by: Simon Glass <sjg@chromium.org>