rockchip: Use the same boot_targets for all boards
It doesn't really matter if we mention things which are not present. For
example, if 'nvme' is included but the board does not support it, it just
continues with the next item in the list.

It is simpler to use the same target list for all boards, so drop the
different one for rk3399.

Series-changes: 4
- Rebase to -next

Series-changes: 4
- Add new patch to use the same boot_targets for all boards

Series-changes: 5
- Drop patch to relax the argument requirements for bootflow scan

Change-Id: I3dd639e8c7824c9b0301884a1180cf3b9107766e
Signed-off-by: Simon Glass <sjg@chromium.org>