bootstd: Support booting EFI where multiple options exist
The current EFI implementation has a strange quirk where it watches
loaded files and uses the last-loaded file to determine the device that
is being booted from.

This is confusing with bootstd, where multiple options may exist. Even
loading a device tree will cause it to go wrong. There is no API for
passing this information, since the only entry into booting an EFI image
is the 'bootefi' command.

To work around this, call efi_set_bootdev() for EFI images, if possible,
just before booting.

Series-changes: 5
- Add new patch to support booting EFI where multiple options exist

Change-Id: Ifa423a8f295b3c11e50821222b0db1e869d0c051
Signed-off-by: Simon Glass <sjg@chromium.org>