bootstd: Report missing labels only when asked
Use the -l flag to indicate whether to report missing uclasses.

Also try to be more helpful when no devices are found. For example, when
we see something 'scsi0' requested and nothing was found, this indicates
that there are no SCSI devices, so show a suitable message.

Move messages out of the low-level functions so that silent operation
is possible.

This means they are never reported unless BOOTSTD_FULL is enabled, since
the -l flag cannot otherwise be set.

Suggested-by: Tom Rini <trini@konsulko.com>

Series-changes: 6
- Add new patch to report missing labels only when asked

Change-Id: I34e1ec50c59a5a2427c692f612cd2cfa8d30d325
Signed-off-by: Simon Glass <sjg@chromium.org>