bootstd: Adjust code ordering to work around compiler quirk
At present when debugging is off, bootdev_find_in_blk() sometimes fails
to find a valid bootflow, e.g. with virtio. Accessing the 'blk' variable
later in the function seems to correct it.

Move the 'ret' check before the debug statement and set the block device
again aftewards, to work around this.

Series-changes: 8
- Add new patch to adjust code ordering to work around compiler quirk

Change-Id: Id1d2749652f1e5f4cb07d1ee6f2f71983086b308
Signed-off-by: Simon Glass <sjg@chromium.org>