CI: Enable sandbox build for Windows
Add a new rule to build sandbox for Windows. For now, no tests are run in
this configuration.

Series-to: u-boot
Series-cc: bin
Cover-letter:
Allow building sandbox with MSYS2
This expands the existing work to allow sandbox to build and run on
Windows using MSYS2.

It also fixes a few issues so that binman can be used.

There are various limitations and some features do not work fully yet. In
particular, weak functions are not well supported on Windows so these are
disabled. Various minor compiler-flag and filename adjustments are also
needed.

This does not support SDL, so U-Boot has no display. This is potentially
possible, but requires using minggw instead of the base toolchain [1] so
needs to be dealt with separately.

[1] https://gist.github.com/thales17/fb2e4cff60890a51d9dddd4c6e832ad2
END

Signed-off-by: Simon Glass <sjg@chromium.org>