binman: Provide a way to specific the fdt-list directly
Sometimes multiple boards are built with binman and it is useful to
specify a different FDT list for each. At present this is not possible
without providing multiple values of the of-list entryarg (which is not
supported in the U-Boot build system).

Allow a fit,fdt-list-val string-list property to be used instead.

Series-to: u-boot
Series-cc: Jan Kiszka <jan.kiszka@siemens.com>

Signed-off-by: Simon Glass <sjg@chromium.org>