bootstd: Init the size before reading extlinux file
The implementation in extlinux_pxe_getfile() does not pass a valid size
to bootmeth_read_file(), so this can fail if the uninited value happens to
be too small.

Fix this.

Series-to: u-boot
Series-cc: trini
Cover-letter:
bootstd: Fix some reported problems
This little series fixes a few problems reported over the past few
months.

The main changes is to the EFI bootmeth, which now delays reading the .efi
file until it is actually needed.
END

Signed-off-by: Simon Glass <sjg@chromium.org>