x86: coreboot: Enable support for CBFS
This is normally used with coreboot, so enable support for it in the
coreboot builds.

Add an example to show how it is used.

Series-to: u-boot
Series-cc: bin
Series-cc: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Series-links: 366139
Series-version: 2
Cover-letter:
x86: Improve bootstd support
This series provides some bootstd fixes so that IDE can be used reliably
for booting a distro. It also includes some documentation on how to boot
from coreboot in various scenarios, using QEMU.

With this it is possible to boot Ubuntu 2023.04 from coreboot64, for
example.
END

Signed-off-by: Simon Glass <sjg@chromium.org>