test: capsule: Add files needed for testing EFI capsule updates
Add the files that would be needed for testing the EFI capsule update
functionality. These include the keys needed for signing and
authenticating the capsules with the capsule authentication
functionality enabled. This includes the public key in form of the EFI
Signature List(ESL) file which will be embedded in the platform's DTB.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>