part: Add a new function to detect partition tables
There is currently no exported function to detect whether a partition
table of a particular type is present. The part_get_info_by_type()
function is close, but not quite the same.

This function typically happens when part_init() is called, but for
situations where multiple partitions are used, this selects the
highest-priority one.

Add a new function and make part_get_info_by_type() use it.

Signed-off-by: Simon Glass <sjg@chromium.org>