bootstd: Tidy up some looging statements
Add a hyphen so that the debugging output reads better, in that it shows
it is related to a previous logging statement.

Drop the 'finished uclass' logging statement, since by then dev is NULL.
Instead, show the uclass before finding the next device.

Mention the device at the top of iter_inc_dev().

Signed-off-by: Simon Glass <sjg@chromium.org>