fs: Fix a confusing error about overlapping regions
When lmb runs out of space in its internal tables, it gives errors on
every fs load operation. This is horribly confusing, as the poor user
tries different memory regions one at a time.

Use the updated API to check the error code and print a helpful message.
Also, allow the operation to proceed.

Signed-off-by: Simon Glass <sjg@chromium.org>