efi: Use the installed SMBIOS tables
U-Boot should set up the SMBIOS tables during startup, as it does on x86.
Ensure that it does this correctly on non-x86 machines too, by creating
an event spy for last-stage init.

Tidy up the installation-condition code while we are here. For
mvebu_armada-37xx add a dependcy on DM_MDIO since it seems that
last_stage_init() is only enabled for some boards.

Signed-off-by: Simon Glass <sjg@chromium.org>