Allow Python packages to be dropped
When building in a portage chroot, we do not have the environment needed
to build pylibfdt. It is instead build as a separate package.

Provide a build option to tell U-Boot to skip this part of the build. We
still need it to use binman, etc. but don't need it to build its
dependencies.

BUG=b:297564447
TEST=manual
$ NO_PYTHON=1 crosfw coreboot64 -F
$ ls /tmp/b/coreboot/scripts/dtc/pylibfdt
See that the directory does not exist

Change-Id: I3fe7ac62de70902fd1be093bcda7c53acd43daf8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/u-boot/+/4814578
Reviewed-by: Mike Frysinger <vapier@chromium.org>