bootstd: Add a return code to bootflow menu
Return an error when the user does not select an OS, so we know whether
to boot or not.

Move calling of bootflow_menu_run() into a separate so we can call it
from other places.

Expand the test to cover these cases.

Add some documentation also, while we are here.

Series-changes: 2
- Add new patch to add a return code to bootflow menu

Change-Id: Iaa9b5bbab0f25012cfc68a803a9da83e71b19aaf
Signed-off-by: Simon Glass <sjg@chromium.org>