usb: Avoid unbinding devices in use by bootflows
When a USB device is unbound, it causes any bootflows attached to it to
be removed, via a call to bootdev_clear_bootflows() from
bootdev_pre_unbind(). This obviously makes it impossible to boot the
bootflow.

However, when booting a bootflow that relies on USB, usb_stop() is
called, which unbinds the device. At that point any information
attached to the bootflow is dropped.

This is quite risky since the contents of freed memory are not
guaranteed to remain unchanged. Depending on what other options are
done before boot, a hard-to-find bug may crop up.

There is no need to unbind all the USB devices just to quiesce them.
Add a new usb_pause() call which removes them but leaves them bound.

Series-changes: 2
- Add new patch to avoid unbinding devices in use by bootflows

Series-changes: 4
- Don't rename the legacy-USB functions
- Add a bit more detail to the comment

Series-changes: 5
- Adjust motivation for patch, since the EFI hang is fixed

Patch-cc: Mattijs Korpershoek <mkorpershoek@baylibre.com>
Tested-by: Shantur Rathore <i@shantur.com>

Change-Id: If206027372f73ce32480223e5626f4b944e281b7
Signed-off-by: Simon Glass <sjg@chromium.org>